LF-5127 Extract the Radar Chart and related logic from TapeResults
